backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
alembic==1.16.5
psycopg[binary]==3.2.10
pydantic==2.11.7
pydantic-settings==2.10.1
python-jose==3.5.0
passlib==1.7.4
python-multipart==0.0.20
email-validator==2.2.0
pymupdf==1.27.2

boto3>=1.34,<2
